+++ lib/client/client.ts
@@ -28,6 +28,7 @@ import { SingleConnection } from "./connection";
 import { ClientError } from "./errors";
 import { deleteMsg } from "../operations/deleteMsg";
 import { ConnectionPool } from "../mixins/connection-pool";
+import { announce } from "../operations/announce";
 
 const log = debugLogger("dank-twitch-irc:client");
 
@@ -236,6 +237,13 @@ export class ChatClient extends BaseClient {
     messageID: string
   ): Promise<void> {
     await deleteMsg(this.requireConnection(), channelName, messageID);
+  }
+
+  public async announce(
+    channelName: string,
+    messageText: string
+  ): Promise<void> {
+    await announce(this.requireConnection(), channelName, messageText);
   }
 
   public async whisper(username: string, message: string): Promise<void> {